Create birdar.py
Working on re-creating BIRDAR in Python 2.7.3 because of recent requests. 
PHP works great on a server, but most people don't use it locally- so this
is an effort to make it easier for those users.

So far all this script does is download the radar imagery. You will need to set up 
a home directory with folders for each radar you want to collect
data from (ex. a folder named "KDOX" for Dover, DE) and within each radar
folder you need to create two product folders named "bref1" and "vel1" for reflectivity
and velocity, respectively. When you run the script, Python will look in your home dir and note
all of the folders starting with the letter "K", then it will iterate over them and go out to get
the radar data for each of the stations for both of the products and put them in the appropriate
folders. 

I'm working on getting ImageMagick to work from within Python and will update when I do. Anyone else who
wants to contribute is more than welcome!!!

Cheers

David